backend/requirements-dev.txt
# filepath: c:\Users\ismai\OneDrive\Bureau\TrelloClone\backend\requirements-dev.txt
# Development and testing dependencies
pytest==8.3.5
pytest-django==4.11.1
pytest-cov==6.0.0
factory-boy==3.3.3
coverage==7.8.2
black==25.1.0
flake8==7.2.0
isort==6.0.1
model-bakery==1.19.5
freezegun==1.2.1
django-extensions==3.2.3
django-filter==25.1
